requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
bcc==0.18.0
blinker==1.8.2
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.1.7
click-default-group==1.2.4
cloup==2.1.2
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
databases==0.9.0
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
distlib==0.3.8
distro==1.7.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
form==0.0.1
glcontext==2.5.0
greenlet==3.0.3
h11==0.14.0
hidpidaemon==18.4.6
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.29.4
ipython==8.23.0
isosurfaces==0.1.0
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jupyter_client==8.6.1
jupyter_core==5.7.2
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
manim==0.18.0.post0
ManimPango==0.5.0
mapbox-earcut==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
moderngl==5.10.0
moderngl-window==2.4.4
more-itertools==8.10.0
mount==1.0.0
multipledispatch==1.0.0
mysql-connector-python==9.0.0
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.0
orjson==3.10.6
packaging==24.0
parso==0.8.4
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.2.0
pop-transition==1.1.2
prompt-toolkit==3.0.43
protobuf==3.12.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.26.0
pycparser==2.22
pycups==2.0.1
pydantic==2.8.2
pydantic_core==2.20.1
pydbus==0.6.0
pydub==0.25.1
pyglet==2.1.dev1
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrr==0.10.3
python-apt==2.4.0+ubuntu1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-gnupg==0.4.8
python-multipart==0.0.9
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.0.2
repolib==2.2.1
repoman==1.4.0
requests==2.31.0
rich==13.7.1
scipy==1.13.0
screeninfo==0.8.1
SecretStorage==3.3.1
sessioninstaller==0.0.0
shellingham==1.5.4
six==1.16.0
skia-pathops==0.7.4
sniffio==1.3.1
SQLAlchemy==2.0.31
srt==3.5.3
stack-data==0.6.3
starlette==0.37.2
svgelements==1.9.6
svgwrite==1.4.3
systemd-python==234
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
Tree==0.2.4
typer==0.12.3
typing_extensions==4.11.0
u==0.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.10.0
urllib3==1.26.5
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.26.3
wadllib==1.3.6
watchdog==3.0.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
xdg==5
xkit==0.0.0
zipp==1.0.0
